vantage-ai/veris/requirements.txt
# Agent orchestration
crewai>=0.80.0
crewai-tools>=0.14.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.27.0

# SEC EDGAR — financial filings & XBRL financials
edgartools>=4.0.0

# OpenAI SDK (used by CrewAI internally)
openai>=1.30.0

# Numeric
numpy>=1.26.0
